+++ sh/audit.performance.sh
@@ -1,1 +1,8 @@
-npx lighthouse-multi --depth=100 --endpoints=https://www.congressthecardgame.com
+DIRECTORY=$(date +"%Y-%m-%d")
+if [ ! -d "$DIRECTORY" ]; then
+  mkdir ./reports/"$DIRECTORY"
+fi
+npx lighthouse-multi \
+  --depth=100 \
+  --endpoints=https://www.congressthecardgame.com \
+  --output=reports/"$DIRECTORY"