+++ sh/audit.sh
@@ -2,7 +2,7 @@ DIRECTORY=$(date +"%Y-%m-%d")
 if [ ! -d "$DIRECTORY" ]; then
   mkdir ./reports/"$DIRECTORY"
 fi
-  lighthouse https://www.congressthecardgame.com \
+lighthouse https://www.congressthecardgame.com \
   --emulated-form-factor=none \
   --output=html \
   --output-path=./reports/"$DIRECTORY"/report-$(date +"%Y-%m-%d-%I-%M-%S-%p").html \